// src/canisters/nft/nft_ogy/module_47a7c018/ogy_47a7c018.did.d.ts
export type CandyShared =
    | { Int: bigint }
    | { Map: Array<[CandyShared, CandyShared]> }
    | { Nat: bigint }
    | { Set: Array<CandyShared> }
    | { Nat16: number }
    | { Nat32: number }
    | { Nat64: bigint }
    | { Blob: Array<number> }
    | { Bool: boolean }
    | { Int8: number }
    | { Ints: Array<bigint> }
    | { Nat8: number }
    | { Nats: Array<bigint> }
    | { Text: string }
    | { Bytes: Array<number> }
    | { Int16: number }
    | { Int32: number }
    | { Int64: bigint }
    | { Option: [] | [CandyShared] }
    | { Floats: Array<number> }
    | { Float: number }
    | { Principal: Principal }
    | { Array: Array<CandyShared> }
    | { Class: Array<PropertyShared> };
// ...
export interface PropertyShared {
    value: CandyShared;
    name: string;
    immutable: boolean;
}